refactor: comment display to the component library
