+++ packages/components/src/CommentsTitle/CommentsTitle.stories.tsx
@@ -1,0 +1,24 @@
+import { CommentsTitle } from './CommentsTitle'
+
+import type { Meta, StoryFn } from '@storybook/react'
+import type { Comment } from 'oa-shared'
+
+export default {
+  title: 'Commenting/CommentsTitle',
+  component: CommentsTitle,
+} as Meta<typeof CommentsTitle>
+
+export const NoComments: StoryFn<typeof CommentsTitle> = () => (
+  <CommentsTitle comments={[]} />
+)
+
+export const OneComment: StoryFn<typeof CommentsTitle> = () => {
+  const comment = {} as Comment
+
+  return <CommentsTitle comments={[comment]} />
+}
+
+export const MultipleComments: StoryFn<typeof CommentsTitle> = () => {
+  const comment = {} as Comment
+  return <CommentsTitle comments={[comment, comment, comment]} />
+}
