+++ packages/components/src/DiscussionContainer/DiscussionContainer.stories.tsx
@@ -6,7 +6,7 @@ import { DiscussionContainer } from './DiscussionContainer'
 import type { Meta, StoryObj } from '@storybook/react'
 
 export default {
-  title: 'Discussions/DiscussionContainer',
+  title: 'Commenting/DiscussionContainer',
   component: DiscussionContainer,
 } as Meta<typeof DiscussionContainer>
