+++ packages/components/src/index.ts
@@ -12,6 +12,8 @@ export { CardProfile } from './CardProfile/CardProfile'
 export { Category } from './Category/Category'
 export { CategoryVerticalList } from './CategoryVerticalList/CategoryVerticalList'
 export { ContentStatistics } from './ContentStatistics/ContentStatistics'
+export { CommentDisplay } from './CommentDisplay/CommentDisplay'
+export { CommentsTitle } from './CommentsTitle/CommentsTitle'
 export { CommentItem } from './CommentItem/CommentItem'
 export { CommentList } from './CommentList/CommentList'
 export { CreateReply } from './CreateReply/CreateReply'
