+++ src/services/notificationsService.server.ts
@@ -6,8 +6,12 @@ import { randomID } from 'src/utils/helpers'
 import { userService } from './userService.server'
 
 import type { SupabaseClient } from '@supabase/supabase-js'
-import type { INotification, IUserDB, NotificationType } from 'oa-shared'
-import type { DBComment } from 'src/models/comment.model'
+import type {
+  DBComment,
+  INotification,
+  IUserDB,
+  NotificationType,
+} from 'oa-shared'
 import type { DBProfile } from 'src/models/profile.model'
 import type { DBQuestion } from 'src/models/question.model'
